requirements.txt
# PDF 텍스트 추출
pdfplumber>=0.9.0
PyPDF2>=3.0.1

# LLM API 연동
openai>=1.3.0
anthropic>=0.7.0

# 환경 변수 관리
python-dotenv>=1.0.0

# CLI 인터페이스
click>=8.1.0

# Web UI
streamlit>=1.28.0

# 텍스트 처리 및 유틸리티
requests>=2.31.0

# 패키징 (개발용)
pyinstaller>=5.13.0